Add Test suffix to test classes.
