+++ api/src/test/java/bio/terra/tanagra/aousynthetic/ConditionEntityQueriesTest.java
@@ -1,9 +1,9 @@
 package bio.terra.tanagra.aousynthetic;
 
-import static bio.terra.tanagra.aousynthetic.UnderlayUtils.ALL_VISIT_ATTRIBUTES;
+import static bio.terra.tanagra.aousynthetic.UnderlayUtils.ALL_CONDITION_ATTRIBUTES;
 import static bio.terra.tanagra.aousynthetic.UnderlayUtils.BQ_DATASET_SQL_REFERENCE;
+import static bio.terra.tanagra.aousynthetic.UnderlayUtils.CONDITION_ENTITY;
 import static bio.terra.tanagra.aousynthetic.UnderlayUtils.UNDERLAY_NAME;
-import static bio.terra.tanagra.aousynthetic.UnderlayUtils.VISIT_ENTITY;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import bio.terra.tanagra.app.controller.EntityInstancesApiController;
@@ -20,34 +20,40 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 /**
- * Tests for visit entity queries on the AoU synthetic underlay. There is no need to specify an
+ * Tests for condition entity queries on the AoU synthetic underlay. There is no need to specify an
  * active profile for this test, because we want to test the main application definition.
  */
-public class VisitEntityQueries extends BaseSpringUnitTest {
-  private static final Logger LOG = LoggerFactory.getLogger(VisitEntityQueries.class);
+public class ConditionEntityQueriesTest extends BaseSpringUnitTest {
+  private static final Logger LOG = LoggerFactory.getLogger(ConditionEntityQueriesTest.class);
   @Autowired private EntityInstancesApiController apiController;
 
   @Test
-  @DisplayName("correct SQL string for listing all visit entity instances")
-  void generateSqlForAllVisitEntities() {
+  @DisplayName("correct SQL string for listing all condition entity instances")
+  void generateSqlForAllConditionEntities() {
     ResponseEntity<ApiSqlQuery> response =
         apiController.generateDatasetSqlQuery(
             UNDERLAY_NAME,
-            VISIT_ENTITY,
+            CONDITION_ENTITY,
             new ApiGenerateDatasetSqlQueryRequest()
                 .entityDataset(
                     new ApiEntityDataset()
-                        .entityVariable("visit_alias")
-                        .selectedAttributes(ALL_VISIT_ATTRIBUTES)));
+                        .entityVariable("condition_alias")
+                        .selectedAttributes(ALL_CONDITION_ATTRIBUTES)));
     assertEquals(HttpStatus.OK, response.getStatusCode());
     String generatedSql = response.getBody().getQuery();
     LOG.info(generatedSql);
     assertEquals(
-        "SELECT visit_alias.concept_id AS concept_id, "
-            + "visit_alias.concept_name AS concept_name "
+        "SELECT condition_alias.concept_id AS concept_id, "
+            + "condition_alias.concept_name AS concept_name, "
+            + "condition_alias.vocabulary_id AS vocabulary_id, "
+            + "(SELECT vocabulary.vocabulary_name FROM `"
+            + BQ_DATASET_SQL_REFERENCE
+            + "`.vocabulary WHERE vocabulary.vocabulary_id = condition_alias.vocabulary_id) AS vocabulary_name, "
+            + "condition_alias.standard_concept AS standard_concept, "
+            + "condition_alias.concept_code AS concept_code "
             + "FROM (SELECT * FROM `"
             + BQ_DATASET_SQL_REFERENCE
-            + "`.concept WHERE vocabulary_id = 'Visit') AS visit_alias "
+            + "`.concept WHERE domain_id = 'Condition') AS condition_alias "
             + "WHERE TRUE",
         generatedSql);
   }
