+++ api/src/test/java/bio/terra/tanagra/aousynthetic/IngredientEntityQueriesTest.java
@@ -1,8 +1,8 @@
 package bio.terra.tanagra.aousynthetic;
 
-import static bio.terra.tanagra.aousynthetic.UnderlayUtils.ALL_CONDITION_ATTRIBUTES;
+import static bio.terra.tanagra.aousynthetic.UnderlayUtils.ALL_INGREDIENT_ATTRIBUTES;
 import static bio.terra.tanagra.aousynthetic.UnderlayUtils.BQ_DATASET_SQL_REFERENCE;
-import static bio.terra.tanagra.aousynthetic.UnderlayUtils.CONDITION_ENTITY;
+import static bio.terra.tanagra.aousynthetic.UnderlayUtils.INGREDIENT_ENTITY;
 import static bio.terra.tanagra.aousynthetic.UnderlayUtils.UNDERLAY_NAME;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -20,40 +20,40 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 /**
- * Tests for condition entity queries on the AoU synthetic underlay. There is no need to specify an
+ * Tests for ingredient entity queries on the AoU synthetic underlay. There is no need to specify an
  * active profile for this test, because we want to test the main application definition.
  */
-public class ConditionEntityQueries extends BaseSpringUnitTest {
-  private static final Logger LOG = LoggerFactory.getLogger(ConditionEntityQueries.class);
+public class IngredientEntityQueriesTest extends BaseSpringUnitTest {
+  private static final Logger LOG = LoggerFactory.getLogger(IngredientEntityQueriesTest.class);
   @Autowired private EntityInstancesApiController apiController;
 
   @Test
-  @DisplayName("correct SQL string for listing all condition entity instances")
-  void generateSqlForAllConditionEntities() {
+  @DisplayName("correct SQL string for listing all ingredient entity instances")
+  void generateSqlForAllIngredientEntities() {
     ResponseEntity<ApiSqlQuery> response =
         apiController.generateDatasetSqlQuery(
             UNDERLAY_NAME,
-            CONDITION_ENTITY,
+            INGREDIENT_ENTITY,
             new ApiGenerateDatasetSqlQueryRequest()
                 .entityDataset(
                     new ApiEntityDataset()
-                        .entityVariable("condition_alias")
-                        .selectedAttributes(ALL_CONDITION_ATTRIBUTES)));
+                        .entityVariable("ingredient_alias")
+                        .selectedAttributes(ALL_INGREDIENT_ATTRIBUTES)));
     assertEquals(HttpStatus.OK, response.getStatusCode());
     String generatedSql = response.getBody().getQuery();
     LOG.info(generatedSql);
     assertEquals(
-        "SELECT condition_alias.concept_id AS concept_id, "
-            + "condition_alias.concept_name AS concept_name, "
-            + "condition_alias.vocabulary_id AS vocabulary_id, "
+        "SELECT ingredient_alias.concept_id AS concept_id, "
+            + "ingredient_alias.concept_name AS concept_name, "
+            + "ingredient_alias.vocabulary_id AS vocabulary_id, "
             + "(SELECT vocabulary.vocabulary_name FROM `"
             + BQ_DATASET_SQL_REFERENCE
-            + "`.vocabulary WHERE vocabulary.vocabulary_id = condition_alias.vocabulary_id) AS vocabulary_name, "
-            + "condition_alias.standard_concept AS standard_concept, "
-            + "condition_alias.concept_code AS concept_code "
+            + "`.vocabulary WHERE vocabulary.vocabulary_id = ingredient_alias.vocabulary_id) AS vocabulary_name, "
+            + "ingredient_alias.standard_concept AS standard_concept, "
+            + "ingredient_alias.concept_code AS concept_code "
             + "FROM (SELECT * FROM `"
             + BQ_DATASET_SQL_REFERENCE
-            + "`.concept WHERE domain_id = 'Condition') AS condition_alias "
+            + "`.concept WHERE concept_class_id = 'Ingredient') AS ingredient_alias "
             + "WHERE TRUE",
         generatedSql);
   }
