+++ api/src/test/java/bio/terra/tanagra/aousynthetic/VisitEntityQueriesTest.java
@@ -1,9 +1,9 @@
 package bio.terra.tanagra.aousynthetic;
 
-import static bio.terra.tanagra.aousynthetic.UnderlayUtils.ALL_DEVICE_ATTRIBUTES;
+import static bio.terra.tanagra.aousynthetic.UnderlayUtils.ALL_VISIT_ATTRIBUTES;
 import static bio.terra.tanagra.aousynthetic.UnderlayUtils.BQ_DATASET_SQL_REFERENCE;
-import static bio.terra.tanagra.aousynthetic.UnderlayUtils.DEVICE_ENTITY;
 import static bio.terra.tanagra.aousynthetic.UnderlayUtils.UNDERLAY_NAME;
+import static bio.terra.tanagra.aousynthetic.UnderlayUtils.VISIT_ENTITY;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import bio.terra.tanagra.app.controller.EntityInstancesApiController;
@@ -20,40 +20,34 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 /**
- * Tests for device entity queries on the AoU synthetic underlay. There is no need to specify an
+ * Tests for visit entity queries on the AoU synthetic underlay. There is no need to specify an
  * active profile for this test, because we want to test the main application definition.
  */
-public class DeviceEntityQueries extends BaseSpringUnitTest {
-  private static final Logger LOG = LoggerFactory.getLogger(DeviceEntityQueries.class);
+public class VisitEntityQueriesTest extends BaseSpringUnitTest {
+  private static final Logger LOG = LoggerFactory.getLogger(VisitEntityQueriesTest.class);
   @Autowired private EntityInstancesApiController apiController;
 
   @Test
-  @DisplayName("correct SQL string for listing all device entity instances")
-  void generateSqlForAllDeviceEntities() {
+  @DisplayName("correct SQL string for listing all visit entity instances")
+  void generateSqlForAllVisitEntities() {
     ResponseEntity<ApiSqlQuery> response =
         apiController.generateDatasetSqlQuery(
             UNDERLAY_NAME,
-            DEVICE_ENTITY,
+            VISIT_ENTITY,
             new ApiGenerateDatasetSqlQueryRequest()
                 .entityDataset(
                     new ApiEntityDataset()
-                        .entityVariable("device_alias")
-                        .selectedAttributes(ALL_DEVICE_ATTRIBUTES)));
+                        .entityVariable("visit_alias")
+                        .selectedAttributes(ALL_VISIT_ATTRIBUTES)));
     assertEquals(HttpStatus.OK, response.getStatusCode());
     String generatedSql = response.getBody().getQuery();
     LOG.info(generatedSql);
     assertEquals(
-        "SELECT device_alias.concept_id AS concept_id, "
-            + "device_alias.concept_name AS concept_name, "
-            + "device_alias.vocabulary_id AS vocabulary_id, "
-            + "(SELECT vocabulary.vocabulary_name FROM `"
-            + BQ_DATASET_SQL_REFERENCE
-            + "`.vocabulary WHERE vocabulary.vocabulary_id = device_alias.vocabulary_id) AS vocabulary_name, "
-            + "device_alias.standard_concept AS standard_concept, "
-            + "device_alias.concept_code AS concept_code "
+        "SELECT visit_alias.concept_id AS concept_id, "
+            + "visit_alias.concept_name AS concept_name "
             + "FROM (SELECT * FROM `"
             + BQ_DATASET_SQL_REFERENCE
-            + "`.concept WHERE domain_id = 'Device') AS device_alias "
+            + "`.concept WHERE vocabulary_id = 'Visit') AS visit_alias "
             + "WHERE TRUE",
         generatedSql);
   }
